megacommit: xcode project, silence some warnings, more reorg
* replace some comma operators that were causing warnings

* use enum for vt_line, vt_ccw_arc, vt_cw_arc; as part of trying to
unravel a pocketing problem.  Ultimately it was due to floating point
errors in reading DXF entities.

*

+++ dxf.cpp
@@ -1359,7 +1359,7 @@ void CDxfRead::get_line()
 	m_ifs->getline(m_str, 1024);
 
 	char str[1024];
-	int len = strlen(m_str);
+	size_t len = strlen(m_str);
 	int j = 0;
 	bool non_white_found = false;
 	for(int i = 0; i<len; i++){
